Fix package installation problems and refactor classes
